backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.3
django-filter==23.5
djangorestframework==3.12.4
djoser==2.1.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3.post1
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.9.0
urllib3==1.26.18
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
python-dotenv==1.0.1
